airflow/requirements.txt
apache-airflow-providers-google
pyarrow
boto3
wget